pyproject.toml
[project]
name = "proposal-agent"
version = "0.3.0"
description = "LangGraph Multi-Agent 기반 용역 제안서 자동 생성 AI 에이전트"
requires-python = ">=3.11"
dependencies = [
    # Core LangGraph & LangChain
    "langgraph>=0.2.0",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-anthropic>=0.2.0",
    # FastAPI
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # Anthropic Claude API
    "anthropic>=0.40.0",
    # Utilities
    "python-multipart>=0.0.12",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    # Document processing
    "pypdf2>=3.0.0",
    "python-docx>=1.1.0",
    "python-pptx>=1.0.0",
    # Optional: Vector DB support
    "chromadb>=0.5.0",
    # Supabase client
    "supabase>=2.0.0",
    # HTTP 클라이언트 (나라장터 API)
    "aiohttp>=3.9.0",
    "python-hwpx>=2.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "state", "graph", "tools", "prompts", "config"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
